#8: moved previous files which are maybe not needed anymore to "_old" directory
