+++ _old/go/internal.go
