+++ _old/main.go
@@ -1,4 +1,4 @@
-package src
+package _old
 
 //import (
 	// WARNING!
